EOS-28614: remove openldap dependency
