+++ py-utils/src/utils/data/db/openldap/__init__.py
@@ -15,4 +15,4 @@
 # For any questions about this software or licensing,
 # please email opensource@seagate.com or cortx-questions@seagate.com.
 
-from cortx.utils.data.db.openldap.storage import OpenLdap
+from cortx.utils.data.db.openldap.openldap import OpenLdap
